Replace int with double

+++ src/calculator/Controller.java
@@ -19,44 +19,44 @@ public class Controller {
     @FXML public Button btnFloorDivision;
     @FXML public Button btnReset;
 
-    private int value1;
-    private int value2;
+    private double value1;
+    private double value2;
 
     private void updateValues() {
         // For more efficiency, the values are only updated once one of the operand buttons is pressed.
-        value1 = Integer.parseInt(fieldValue1.getText());
-        value2 = Integer.parseInt(fieldValue2.getText());
+        value1 = Double.parseDouble(fieldValue1.getText());
+        value2 = Double.parseDouble(fieldValue2.getText());
     }
 
     public void btnAdditionOnAction(ActionEvent actionEvent) {
         updateValues();
-        int result = value1 + value2;
+        double result = value1 + value2;
         fieldResult.setText(String.valueOf(result));
     }
 
 
     public void btnSubtractionOnAction(ActionEvent actionEvent) {
         updateValues();
-        int result = value1 - value2;
+        double result = value1 - value2;
         fieldResult.setText(String.valueOf(result));
     }
 
     public void btnMultiplicationOnAction(ActionEvent actionEvent) {
         updateValues();
-        int result = value1 * value2;
+        double result = value1 * value2;
         fieldResult.setText(String.valueOf(result));
     }
 
     public void btnDivisionOnAction(ActionEvent actionEvent) {
         updateValues();
-        int result = value1 / value2;
+        double result = value1 / value2;
         fieldResult.setText(String.valueOf(result));
     }
 
     public void btnPowerOnAction(ActionEvent actionEvent) {
         updateValues();
 
-        int result = value1;
+        double result = value1;
         for (int i = 0; i < value2 - 1; i++) {
             result *= value1;
         }
@@ -65,13 +65,13 @@ public class Controller {
 
     public void btnModuloOnAction(ActionEvent actionEvent) {
         updateValues();
-        int result = value1 % value2;
+        double result = value1 % value2;
         fieldResult.setText(String.valueOf(result));
     }
 
     public void btnFloorDivisionOnAction(ActionEvent actionEvent) {
         updateValues();
-        int result = value1 + value2;
+        double result = value1 + value2;
         fieldResult.setText(String.valueOf(result));
     }
